refactor: reduce external visibility
