+++ src/queries/displayValue.ts
@@ -11,7 +11,7 @@ import type {
   QueryAllByQuery,
   QueryByQuery,
 } from './makeQueries';
-import type { TextMatchOptions } from './byText';
+import type { TextMatchOptions } from './text';
 
 const getTextInputNodeByDisplayValue = (
   node: ReactTestInstance,
