+++ src/queries/placeholderText.ts
@@ -14,7 +14,7 @@ import type {
   QueryAllByQuery,
   QueryByQuery,
 } from './makeQueries';
-import type { TextMatchOptions } from './byText';
+import type { TextMatchOptions } from './text';
 
 const getTextInputNodeByPlaceholderText = (
   node: ReactTestInstance,
